framework/common/org.gumtree.core/src/org/gumtree/service/notification/support/LogNotification.java
/*******************************************************************************
 * Copyright (c) 2012 Australian Nuclear Science and Technology Organisation.
 * All rights reserved. This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v1.0
 * which accompanies this distribution, and is available at
 * http://www.eclipse.org/legal/epl-v10.html
 *
 * Contributors:
 *     Bragg Institute - initial API and implementation
 ******************************************************************************/

package org.gumtree.service.notification.support;

public class LogNotification extends AbstractNotification {

	private static final long serialVersionUID = 202304805346879884L;

	private static final String PROTOCOL = "log";

	protected static final String EMPTY_STRING = "";

	private String message;

	public LogNotification() {
		this(EMPTY_STRING);
	}

	public LogNotification(String message) {
		super();
		message = (message == null) ? EMPTY_STRING : message;
		setMessage(message);
	}

	public String getMessage() {
		return message;
	}

	public void setMessage(String message) {
		this.message = message;
	}

	public String getProtocol() {
		return PROTOCOL;
	}

	@Override
	public int hashCode() {
		final int prime = 31;
		int result = super.hashCode();
		result = prime * result + ((message == null) ? 0 : message.hashCode());
		return result;
	}

	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (!super.equals(obj))
			return false;
		if (getClass() != obj.getClass())
			return false;
		LogNotification other = (LogNotification) obj;
		if (message == null) {
			if (other.message != null)
				return false;
		} else if (!message.equals(other.message))
			return false;
		return true;
	}

	@Override
	public String toString() {
		return getToStringHelper().add("message", getMessage()).toString();
	}

}
